fix: work around for loosing stack params in fn
The parameters passed to swpan_task will be lost when the stack is
modified, if not already by the other execution path exiting. There is
therefore no way to pass the kernel parameter on. Instead, we can use
the kill_curr_task function in os.h, which injects the proper kernel
pointer.